RDKB-7153 : Develop Code - Add ATM Tr181 into wifi agent
Reason for change: Develop Tr181 for ATM feature.
This is the first patch of code, few items are left as TBD
to be completed in the next patches.

Test Procedure: Query dmcli eRT getv Device.WiFi.X_RDKCENTRAL-COM_ATM.
The details of the expected output are mentioned in the JIRA.

Risks: None

Change-Id: I6a92a8730e1d8ec397ee5cb48a46786b0a65d998
Signed-off-by: Roopali_Galgali@comcast.com