RDKB-9462:Wifiagent log showing multiple PSM_Get_Record_value2 error
Reason for change: This issue is seen when the device is in factory reset state and Prefer parameter is not available in bbhm.
Whenever a client is connected to private ssid, we check for the status(PSM_Get_Record_Value2 for PreferPrivate) of PreferPrivate feature.
If PreferPrivate parameter is not available in PSM database then we print Warning message and return PreferPrivate value as false.
The fix is to remove warning message from code.

Test Procedure:
Do a factory reset. Enable Xfinity hotspot. Connect a device to private ssid.
Verify AtomConsole log for "CosaDmlWiFi_GetPreferPrivatePsmData - PSM_Get_Record_Value2 returned error - 102" message

Risks: None

Signed-off-by: Raja Shah <raja_shah2@cable.comcast.com>

Change-Id: Ie1443da4c0dbd550aa573cce65ad721aa3532408