RDKB-13410:With preferPrivate enabled, xfinitywifi macfilter is still
Reason for change:After we enable preferPrivate, the PSM DB is updated with proper values, but the
local structure was not updated. Because of which we see macfilter is set to Allow-All.
So after setting the preferprivate to true, when we call the get X_COMCAST-COM_MAC_FilteringMode
for the appropriate ath interfaces, the local structure should be updated with proper values.
Test Procedure:After flashing the image, enable the preferPrivate and check the
X_COMCAST-COM_MAC_FilteringMode for 5,6,9,10
Risks: None
Signed-off-by: Sudharsan Raman <sudharsan_raman@comcast.com>

Change-Id: I2f0629643711ee0a2f88551409074c96695e4192