RDKB-14318,ARRISXB3-6730 : lighttpd fails to launch sometimes
Reason for change: Changing the port numbers 51515 and 58081 as there is a possibility for these ports to be
allocated to some other process as they fall in the local port range which OS can allocate for random ports.

Test Procedure: "Lighttpd is not running" prints should not be present in Arm and Atom.

Risks: None
Signed-off-by: Dhanya_Ravindran2@comcast.com

Change-Id: I8bc603e9da40e7bc6251da195c73083ed4f887a6