TCCBR-2517: Check to see if we are changing the default SSID on CBR
Reason for change: Chaning the default SSID is only checked as part
of the captive portal code. The BCI devices do not have captive
portal mode so to print this log out, we need to check where we
normally set the SSID values.

Test Procedure: Verify that the code checks the default ssid

Risks: Low

Change-Id: I02a54206fffbfc9ca4bebb5509a97de52bb5f94e
Signed-off-by: scott_angel@cable.comcast.com