SKYH4-2065 : Harden CcspWifi for corrupted data structures
Reason for change: It was noted that the data structures returned from
wifi_getNeighboringWiFiDiagnosticResult2 and
wifi_getApAssociatedDeviceDiagnosticResult3 can be partially filled in the
event of an error during processing. In these case, we will drop the data
altogether and free up any allocated memory to be safe.

Test Procedure: Verify that CcspWifi does not exhibit any issues during
diagnostic or neighboir data collection.

Risks: Low

Signed-off-by: Lakshminarayanan Shenbagaraj <lakshminarayanan_shenbagaraj@comcast.com>
Change-Id: I7369db840534a9c9fd052b9c27328b256b0390e6