REFPLTB-622 : Online / Offline status of connected devices on RDKB
Reason for change: Added the support of inactive/active wifi clients status update
Test Procedure: When you connect the client, you are able to see the below lines
in wifi log.
200609-10:11:39.615664 [mod=WIFI, lvl=WARN] [tid=15480]  CosaDMLWiFi_Send_FullHostDetails_To_LMLite-17331 [12:ae:a9:a8:3b:08,Device.WiFi.AccessPoint.1.AssociatedDevice.1,Device.WiFi.SSID.1,-65,1]
200609-10:11:39.618651 [mod=WIFI, lvl=WARN] [tid=15480]  CosaDMLWiFi_Send_FullHostDetails_To_LMLite-17331 [16:bc:8a:14:25:24,Device.WiFi.AccessPoint.1.AssociatedDevice.2,Device.WiFi.SSID.1,-67,1]
200609-10:11:39.636104 [mod=WIFI, lvl=WARN] [tid=15480]  CosaDMLWiFi_Send_FullHostDetails_To_LMLite-17331 [38:e6:0a:2b:82:b5,Device.WiFi.AccessPoint.1.AssociatedDevice.3,Device.WiFi.SSID.1,-41,1]
If you disconnect the wifi clients, on that time the last 2 values will be set to Zero
This is the difference between active and inactive clients.
IN WebUI, go to connected Device > Devices and check the status of connected wifi clients
Risks:Low

Signed-off-by: kpandu549 <keerthana_pandurangan@comcast.com>
Change-Id: Ib463c23b966a07cdfd99a87c22427bb46a8e0747