RDKCOM-2312: RDKBDEV-467: fix misuse of "\0"
Reason for change:
"\0" is a pointer to an empty string (ie effectively the same as ""),
and is NOT the same as a nul charactor (ie a char with value 0).

Usage of "\0" is almost always a bug. If a pointer to an empty string
is needed then use "", if a nul character is needed then use 0 or
'\0'.

Test Procedure:
1. Check the LAN, WAN and Basic Sanity test should be sufficient.
Risks: None

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
(cherry picked from commit fbe444c644bd7f1479be17427dc1117a6a260f52)
Change-Id: I790b1c12a959a8939affa5fcd35b7a859190bf77