RDKBDEV-460: fix init of size argument
Reason for change:
Fix init of size argument passed to CosaGetParamValueString()

The value of ulEntryNameLen will be updated by each call to
CosaGetParamValueString() with the size of the string value returned,
therefore it needs to be reset back to the full size of the output
buffer before every call.

Test Procedure:
1. Check the LAN & WAN are working properly
2. A basic Sanity testing

Risks: None.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: I90ad31c6b367cffa6097343cef1d0ce988c15990