RDKB-29399 : XB3 advertising 802.11b rates when configured with 802.11g/n mode
Reason for change: when changing the radio standard the basic data rates for the radio was not changed. Hence called the API to set the basic data rates.
Test Procedure: change the operating standard to g/n mode and check if the management packets are sent with the transmit rates not equal to the 802.11b standard
Risks: Low
Signed-off-by:ManikandaPrabhu_Selvam@comcast.com

(cherry picked from commit a917d594d661bf3b99d7d00e59741c23a9678ef1)
Change-Id: I5fcd550685b70901dab0289af93a845bd4d37075