RDKBDEV-875: CcspWifAgent: drop support for obsolete _COSA_DRG_CNS_
Reason for change:
drop support for obsolete _COSA_DRG_CNS_ and _COSA_DRG_TPG_ platforms

These are legacy Cisco platforms which are not supported or buildable
in RDKB. They predate the introduction of the RDKB HAL APIs so the
code isn't generally useful to keep around, even as a reference.

  for file in $(find . -type f -name '*.[ch]') ; do unifdef -U_COSA_DRG_TPG_ -U_COSA_DRG_CNS_ -o $file $file ; done

Test Procedure: Build and Sanity.
Risks: None.
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: Id581c3ccfacfefe9b5824070681882b0a4c995e7