RDKB-34822 : Throughput dip seen due to frequent scan
Reason for change: wifi_getRadioChannelsInUse API scans all the
frequency bands and hence there is a dip in the throughput while
scanning. This was called every 5seconds by the monitor thread. Since
this call was needed only for Blaster, moved the API to get called at
the time of blaster.
Test Procedure: Load the build and check for throughput
Risks: Low
Signed-off-by:ManikandaPrabhu_Selvam@comcast.com

Change-Id: I7bd8447404e4838f1161f3aa34c05b5894ff6150