RDKB-37774 : MFPConfig can't be configured as Optional for LNF Secure
Reason for change: In Security_Validate() API, the control return FALSE
if the security mode is mixed and the encryption mode is not mixed. But
there is a sync issue between HAL and DML where the encryption mode is
mixed in HAL and it was not populated in DML as it was blocked by XB6
product flag. So whenever there is a change in the security params for
LNF this call return failure. Hence added some code to populate the DML
encryption when the mode is mixed.
Test Procedure: change the MFPConfig value and check in HAL and DML
Risks: Low
Signed-off-by:ManikandaPrabhu_Selvam@comcast.com

(cherry picked from commit ef295e51a6dbecdc807bde70954b15f34fb07e1d)
Change-Id: Ibd901f4e72593d542331be1ab477639e72eec3f1