RDKB-42740 increase wifi monitor timeout
Impacted Platforms:
All RDKB Boxes

Reason for change:
In this issue, the reported stated that the blaster error percentage in
5.2 release train of XB6 is higher when compared to 4.14.
On checking the logs provided by FDR, we could see blaster getting
skipped many times.On further analysis through debug builds, we could
find that the monitor thread is not detecting the pthread signal from
SetActiveMsmtEnable at certain times when the blast is triggered, thus
resulting in the blast not taking place. On further tries, when signal
is detected, the monitor thread detects the current and previous signals
together, with the blasterqueuecount greater than 1.

Hence we are trying to increase wifi monitor timeout

Test Procedure: trigger 40+ blaster and check for blast skip

Risks: Low

Signed-off-by: Harikrishnan_Raman@comcast.com
Change-Id: I98a5bf3a7b6d9adb8be11e4ccecf4f7d87372494