RDKBDEV-1194 CcspWiFiAgent: Implement setting "at least" operating standards
Steps followed :The following combinations are required to be configurable for OperatingStandards:
2.4G: g/n/ax, n/ax, ax
5G: a/n/ac/ax, n/ac/ax, ac/ax, ax
Current implementation in CcspWifiAgent and BCM wifi_hal only supports everything and ax-only, not the other options.
However, BCM driver supports all the required options with the mode_reqd setting.
Extend the behaviour in WifiAgent to allow it to pass a bitmask of pureModes.
This is possible without changing the HAL API; we are just extending our interpretation of standard X only API.
Implement it in bcm wifihal.
Ideally, the beacons should allow to advertise the actual supported standards.
There is however no practical way to do that. However,
what we can do at minimum is to make sure that the lower basic rates are disabled when g resp. a is not allowed.
This can be tested with InSSIDer, which should report n/ax instead of g/n/ax or b/g/n/ax.

Risks: None

Signed-off-by: Mohamed Shaikh <mohamed.shaikh@t-systems.com>
Change-Id: Ic1186036d55894d4f4652050c48a608fe010fe5e

Change-Id: Iac21323b885a7909f79e0af756d0310c791bdca3